+++ chaosaws/__init__.py
@@ -14,7 +14,7 @@ from chaoslib.types import (Configuration, DiscoveredActivities,
                             DiscoveredSystemInfo, Discovery, Secrets)
 from logzero import logger
 
-__version__ = '0.8.0'
+__version__ = '0.9.0'
 __all__ = ["__version__", "discover", "aws_client", "signed_api_call"]